Add files via upload
E-Agrocart computerized approach for better and clear marketing. This website will provide simple 
online platform for farmer and customer to trade agriculture products. The application is maintained 
among 2 types of users: Customer & Farmers. The login id and password must be required to login 
the system. As it is user friendly, it is simple to understand by stakeholders and just asks the user to 
follow step by step operations by giving him options.
The main objective of this project is to build a website that will help farmers to sell their products, 
by providing them with a large online market to sell their produce. This website helps to provide 
technology and services to the farmers and sellers thus helping them to expand their business and 
provide them with a wider market. Hence improving the present farming process and to an extent 
providing knowledge about recent agricultural issues